LPS-55516 Find updated calendar among the received and replace it on available calendars.
What we *really* want is to update the edited calendar. This is a way of doing
it.

However, this is clearly a bad way of doing this: we are returning a lot of
calendars, filtering them and then using the filtered one. What if we could get
only the calendar we need?