LPS-74277 portal-search-web: Use human-readable params in custom range.
If we are going to use custom ranges, we use to date pickers to define the
start date and end date of the range. In the old facet, we used to have many
params (fromYear, fromMonth, toYear etc.) to have the info in the URL. The
result was a large URL:

...&_com_liferay_search_SearchPortlet_fromYear=2018&_com_liferay_search_SearchPortlet_fromMonth=2&...

In the new facet, we want to use more human-readable params:

...&modifiedFrom=2018-3-1&modifiedTo=2018-4-1&...

So in this commit, we change the display builders, portlet etc. to look to the
new parameters and also generate URLs with them. Of course, the new parameters
imply there is a new way of defining if custom range is selected, so we had to
change it too.

As in the last commit, we moved part of the URL creation to the server side, so
we could drop a good amount of JavaScript from the JSP.

https://issues.liferay.com/browse/LPS-74277