LPS-74277 portal-search-web: Modified facet generate range strings from custom ranges.
The search results are filtered by range strings ("[20180101000000 TO
20180201000000]", for example.) Until now, the modified facet either generated
them by "normalizing" named ranges (such as "past-year") or got them prepared,
because the old portlet used to pass the range strings wholy in the URL.

Now that we use readable params (modifiedFrom=2018-01-01 etc.) we have to
generate the range string from them. So we created a new method and changed the
modified facet.

Also, we changed the modified facet display builder to consider that there is a
facet selected, if custom range is available.

https://issues.liferay.com/browse/LPS-74277