LPS-74277 portal-search-web: Test - ensure modified facet behaves and get values from all parameters.
To effectively filter search by both labeled and custom ranges, we need to
change the modified facet. So we wrote tests to provide a safety net, both for
the modified facet itself and to the DateRangeFactory class, that will create
ranges from two values too now.

The exception is testFacetGetModifiedValueFromStartEnd(), which is supposed to
be red: once it passes, we know the facet is minimally prepared to provide
filtering of the results.

Also, we need to ensure that the modified facet display context marks the
custom range option as selected. We then created tests for modified facet
display builder.

https://issues.liferay.com/browse/LPS-74277