LPS-74299 portal-search-web: change portlet to update search by setting document.location.
First, we refactor functions to be testable; remove empty string that causes
"dangling amperstand." This way, we can reuse functions from facet util (that
later are going to be extracted). We also create updateQueryParameters() .

Then we create new function to update URLs; change updateQueryParameters()
behavior.

Also, we intercept form submit. Until now, we intercepted the form submission
when it was done by clicking in the magnifying glass button. Now we are
intercepting by any way the form is submitted.

Finallly, we set current parameterers into search url.

Instead of submitting the form, we change the keyword and add the resulting
query parameters to the destination URL. Then we go to this URL.

https://issues.liferay.com/browse/LPS-74299